--- a/Gems/GradientSignal/Code/Include/GradientSignal/Components/GradientSurfaceDataComponent.h
+++ b/Gems/GradientSignal/Code/Include/GradientSignal/Components/GradientSurfaceDataComponent.h
@@ -72,7 +72,10 @@ namespace GradientSignal
 
         ////////////////////////////////////////////////////////////////////////
         // SurfaceData::SurfaceDataModifierRequestBus
-        void ModifySurfacePoints(SurfaceData::SurfacePointList& surfacePointList) const override;
+        void ModifySurfacePoints(
+            AZStd::span<const AZ::Vector3> positions,
+            AZStd::span<const AZ::EntityId> creatorEntityIds,
+            AZStd::span<SurfaceData::SurfaceTagWeights> weights) const override;
 
         //////////////////////////////////////////////////////////////////////////
         // LmbrCentral::DependencyNotificationBus