--- a/Gems/GradientSignal/Code/Include/GradientSignal/Components/GradientSurfaceDataComponent.h
+++ b/Gems/GradientSignal/Code/Include/GradientSignal/Components/GradientSurfaceDataComponent.h
@@ -57,6 +57,7 @@ namespace GradientSignal
         static void GetProvidedServices(AZ::ComponentDescriptor::DependencyArrayType& services);
         static void GetIncompatibleServices(AZ::ComponentDescriptor::DependencyArrayType& services);
         static void GetRequiredServices(AZ::ComponentDescriptor::DependencyArrayType& services);
+        static void GetDependentServices(AZ::ComponentDescriptor::DependencyArrayType& services);
         static void Reflect(AZ::ReflectContext* context);
 
         GradientSurfaceDataComponent(const GradientSurfaceDataConfig& configuration);