--- a/Gems/SurfaceData/Code/Include/SurfaceData/SurfaceDataModifierRequestBus.h
+++ b/Gems/SurfaceData/Code/Include/SurfaceData/SurfaceDataModifierRequestBus.h
@@ -32,7 +32,10 @@ namespace SurfaceData
         //! allows multiple threads to call
         using MutexType = AZStd::recursive_mutex;
 
-        virtual void ModifySurfacePoints(SurfacePointList& surfacePointList) const = 0;
+        virtual void ModifySurfacePoints(
+            AZStd::span<const AZ::Vector3> positions,
+            AZStd::span<const AZ::EntityId> creatorEntityIds,
+            AZStd::span<SurfaceData::SurfaceTagWeights> weights) const = 0;
     };
 
     typedef AZ::EBus<SurfaceDataModifierRequests> SurfaceDataModifierRequestBus;